Determine presence of metabolic syndrome using updated NCEP ATPIII criteria.

Determines whether an individual has the metabolic syndrome based on the definition of the NCEP ATPIII criteria, updated in 2005 by Grundy et al.

has_metabolic_syndrome_atpiii(
  is_female,
  waist_circumference = NA,
  systolic_blood_pressure = NA,
  diastolic_blood_pressure = NA,
  triglycerides = NA,
  hdl_cholesterol = NA,
  glucose = NA,
  is_fasting_blood_sample = FALSE,
  has_antihypertensive_drug = FALSE,
  has_lipid_drug = FALSE,
  has_glucose_drug = FALSE
)

Arguments

is_female
TRUE if patient is female, FALSE if patient is a man.

waist_circumference
Waist circumference (cm) (NA by default).

systolic_blood_pressure
Systolic blood pressure (mmHg) (NA by default).

diastolic_blood_pressure
Systolic blood pressure (mmHg) (NA by default).

triglycerides
Triglycerides in (mmol/l) (NA by default).

hdl_cholesterol
HDL cholesterol (mmol/l) (NA by default).

glucose
Glucose (mmol/l) (NA by default).

is_fasting_blood_sample
TRUE if fasting blood sample, FALSE (default) if not or unknown.

has_antihypertensive_drug
TRUE if patient is on antihypertensive drug treatment, otherwise FALSE (default).

has_lipid_drug
TRUE if patient is on fibrate, nicotinic acid or statin, otherwise FALSE (default).

has_glucose_drug
TRUE if patient is on drug treatment for elevated glucose, otherwise FALSE (default).

Value

TRUE if patient has the metabolic syndrome, FALSE is not, NA if indetermined.

Caveats

We do include use of a statin as a drug for HDL cholesterol and triglicerides, even though the original statement is unclear about this point (only including niacin and fibrates).

According to the definition, only glucose needs to be from a fasting blood sample (and not triglycerides or HDL cholesterol). The glucose trait is determined in a fail-fast way: if you do not explicitly set is_fasting_blood_sample = TRUE, the glucose measurement is considered to be non-fastening and is ignored.

This function explicitly handles situations of missing data. For example, with blood pressure 120/80 mmHg and antihypertensive medication, a patient is still considered hypertensive. With blood pressure NA/90 mmHg without antihypertensive medication, a patient is still considered hypertensive. A patient with blood pressure NA/NA mmHg and no antihypertensive medication is considered NA for the trait hypertension, but if a patient has NA/NA mmHg but does use antihypertensive medication, the patient is considered to have hypertension (even though measurements are missing).

References

Updated NCEP ATPIII criteria (Grundy et al., 2005): https://doi.org/10.1161/circulationaha.105.169404.